Build step uses context managers
This enables us to set the credential location only while we are working and then unset it when we are done.

Currently only configured for AWS.

In addition, the logic for push and check_registry had to be moved to the auth function. We want to run the context manager every time we get past the first condition in the deploy step.