Support --set-string and avoid numerical image tag bug
When using the helm CLI and `--set`, it cannot know when it receives
it's argument if the value was meant as a string or as an integer if it
receives the string `12334567`.

The helm CLI has opted to have the `--set-string` flag to enforce a set
value to be a string value even if it looks numerical. We should ensure
that we always set image tags and names as strings, or we run into
issues like
https://github.com/earthlab/hub-ops/pull/199#issuecomment-619099629.