Stop repeated checks for parent images
The git log command we use will give us the same
commit hash over and over again if that was the first
time that directory had been committed. This meant we
would repeatedly check for the existence of the same image.

This fixes that by stopping the generator once we
repeat tags